+++ webrtc/ai/OZEngine/parts_classifier/FeatureExtractor.py
@@ -1,0 +1,14 @@
+class FeatureExtractor:
+    def __init__(self):
+        base_model = VGG16(weights='imagenet')
+        # Customize the model to return features from fully-connected layer
+        self.model = Model(inputs=base_model.input, outputs=base_model.get_layer('fc1').output)
+        
+    def extract(self, img):
+        img = img.resize((224, 224))
+        img = img.convert('RGB')
+        x = image.img_to_array(img)
+        x = np.expand_dims(x, axis=0)
+        x = preprocess_input(x)
+        feature = self.model.predict(x)[0]
+        return feature / np.linalg.norm(feature)
