+++ webrtc/ai/OZEngine/parts_classifier/PartsClassifier.py
@@ -1,0 +1,8 @@
+from OZEngine.parts_classifier import FeatureExtractor
+
+
+class PartsClassifier():
+    def __init__(self):
+        self.feature_extractor = FeatureExtractor()
+
+
